added int he ability to set the day alhtouhg setting the year is failing also did other issue of modifying the time off menu item

+++ source/mini_pc/src/headers/graphics.h
@@ -17,7 +17,7 @@ public:
 
     bool BeginDisplay();
     void ClearDisplay();
-    void WriteToScreen(const std::string &str, uint8_t x = 0, uint8_t y = 0, uint8_t ts = 1);
+    void WriteToScreen(const std::string &str, int x = 0, int y = 0, int ts = 1);
     void Display();
 
     void PrintToScreen(const std::string &str);
